move everything to top level contrib dir
