+++ codegen/test/api/things.test.io/v1/sets/sets.go
@@ -5,7 +5,7 @@ package v1sets
 import (
 	. "github.com/solo-io/skv2/codegen/test/api/things.test.io/v1"
 
-	sksets "github.com/solo-io/skv2/pkg/contrib/sets"
+	sksets "github.com/solo-io/skv2/contrib/pkg/sets"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/sets"
 )
