+++ contrib/codegen/templates/sets/sets.gotmpl
@@ -9,7 +9,7 @@ import (
     . "{{ group_import_path $ }}"
     {{- end}}
 
-    sksets "github.com/solo-io/skv2/pkg/contrib/sets"
+    sksets "github.com/solo-io/skv2/contrib/pkg/sets"
     metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
     "k8s.io/apimachinery/pkg/util/sets"
 )
